Moved assembler to a separate file.
[Delivers #57140426]
